Code for Bug #1706
The logic for copying the 2D space into the linear memory of the numpy array was
wrong. The row limit and the colum limt were interchanged, leading to iterating
over colums when moving over rows (was expected) and vice versa.

The bug was not caught in unit testing because, the tests were done using a 5x5
matrix. The interchange of the dimensions was missed because of the square
shape.

Tests have been modified to work with a 6x4 shape




git-svn-id: https://svnsrv.desy.de/desy/mtca4u/PythonBindings/MtcaMappedDevice/trunk@1457 5f601a5c-88b9-47f3-b851-818f6741d00d